Restrict roles

+++ app/controllers/api/v1/users_controller.rb
@@ -23,27 +23,35 @@ class Api::V1::UsersController < ::Api::BaseController
 
     def update
         @user = User.find(params[:id])
-        if @user.update(user_params)
-            render json: @user
+        if @current_user.role === 'super-admin' || @current_user.id === params[:id].to_i
+            if @user.update(user_params)
+                render json: @user
+            else
+                render json: {error_msg: @user.errors}, status: :unprocessable_entity
+            end
         else
-            render json: {error_msg: @user.errors}, status: :unprocessable_entity
+            render json: {error_msg: 'Unauthorised to update this User details'}, status: :unprocessable_entity
         end
     end
 
     def destroy
-      @user = User.find(params[:id])
+        @user = User.find(params[:id])
 
-      if @user.destroy
-        render json: {success_msg: "User successfully deleted"}
-      else
-        render json: {error_msg: "Unable to deleted the User"}
-      end
+        if @current_user.role === 'super-admin' || @current_user.id === params[:id].to_i
+            if @user.destroy
+                render json: {success_msg: "User successfully deleted"}
+            else
+                render json: {error_msg: "Unable to deleted the User"}
+            end
+        else
+            render json: {error_msg: 'Unauthorised to delete this User'}, status: :unprocessable_entity
+        end
     end
 
     private
 
     def user_params
-        params.permit(:name, :phone, :role, :email, :password)
+        params.permit(:name, :phone, :email, :password)
     end
 
     def encode_token(user_id)
